Find archive by name via archive repository
