+++ src/lib/dbSdk/index.ts
@@ -2,6 +2,7 @@ export * from './getArchivistArchiveKeysMongoSdk'
 export * from './getArchivistArchiveMongoSdk'
 export * from './getArchivistBoundWitnessesMongoSdk'
 export * from './getArchivistPayloadMongoSdk'
+export * from './getBaseMongoSdk'
 export * from './getMongoDBValues'
 export * from './storeArchive'
 export * from './UpsertResult'
