+++ src/middleware/Diviner/Archive/ArchiveRepository.ts
@@ -2,4 +2,4 @@ import { XyoArchive } from '@xyo-network/sdk-xyo-client-js'
 
 import { ReadWriteRepository } from '../../../model'
 
-export type ArchiveRepository = ReadWriteRepository<XyoArchive, XyoArchive, string>
+export type ArchiveRepository = ReadWriteRepository<XyoArchive, XyoArchive | null, string>
