+++ src/middleware/Diviner/index.ts
@@ -1,1 +1,2 @@
+export * from './Archive'
 export * from './Archivist'
