+++ src/middleware/index.ts
@@ -1,5 +1,6 @@
 export * from './archiveLocals'
 export * from './auth'
+export * from './Diviner'
 export * from './doc'
 export * from './nodeEnv'
 export * from './rollbar'
