+++ src/server/addDependencies.ts
@@ -1,7 +1,8 @@
 import { Application } from 'express-serve-static-core'
 
-import { XyoSchemaToPayloadProcessorRegistry } from '../middleware'
+import { MongoDBArchiveRepository, XyoSchemaToPayloadProcessorRegistry } from '../middleware'
 
 export const addDependencies = (app: Application) => {
+  app.archiveRepository = new MongoDBArchiveRepository()
   app.payloadProcessorRegistry = new XyoSchemaToPayloadProcessorRegistry(app)
 }
